move kernels to _internal, set up basic sphinx-docs
