+++ docs/.gitignore
@@ -1,0 +1,2 @@
+source/generated
+build
