+++ setup.cfg
@@ -22,4 +22,3 @@ install_requires =
     anndata >= 0.7
     NaiveDE
     h5py
-
